requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
SQLAlchemy>=2.0.36
Werkzeug==3.0.3
python-dotenv==1.0.0
PyJWT==2.8.0
passlib[argon2]==1.7.4
stripe==7.8.0
requests==2.31.0
python-dateutil==2.8.2
email-validator==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9


